+++ tests/test_set_hash.cpp
@@ -20,7 +20,7 @@
 
 #include "hash.hpp"
 #include "random.hpp"
-#include "set_hash_elligator.hpp"
+#include "set_hash.hpp"
 
 #include <iostream>
 #include <vector>
